DF-6367 landscape editors (WIP)
Former-commit-id: 0961d1276f979aad0f6ba50b6e7f4327e621fc4e

+++ Tools/ResourceEditor/Classes/Qt/Scene/System/LandscapeEditorDrawSystem/RulerToolProxy.cpp
@@ -34,7 +34,7 @@ RulerToolProxy::RulerToolProxy(int32 size)
 :	size(size)
 ,	spriteChanged(false)
 {
-	rulerToolTexture = Texture::CreateFBO((float32)size, (float32)size, FORMAT_RGBA8888/*, Texture::DEPTH_NONE*/);
+    rulerToolTexture = Texture::CreateFBO((uint32)size, (uint32)size, FORMAT_RGBA8888/*, Texture::DEPTH_NONE*/);
 #if RHI_COMPLETE_EDITOR
     RenderHelper::Instance()->Set2DRenderTarget(rulerToolTexture);
     RenderManager::Instance()->ClearWithColor(0.f, 0.f, 0.f, 0.f);
